+++ src/main/javascript/Core/ContextObject.js
@@ -2,14 +2,33 @@
 class ContextObject
 {
   /**
+   * @param {Function} propertyProvider
    * @param {String} type the context type
    * @param {String} entityId the id of the Deskpro Entity referenced by this context
    * @param {CustomFieldsClient} [customFields] A client for the `CustomFields` API of this Deskpro Object, if supported
    * @param {...*} rest
    * @constructor
    */
-  constructor({ type, entityId, customFields, ...rest }) {
-    this.props = { type, entityId, customFields, ...rest }
+  constructor({ propertyProvider, type, entityId, customFields, ...rest }) {
+    this.props = { propertyProvider, type, entityId, customFields, ...rest }
+  }
+
+  /**
+   * Retrieves the value of a custom field field from this object is supported
+   *
+   * @param {String} id
+   * @param {*} defaultValue
+   * @return {Promise<String|null|Array<String|null>, Error>}
+   */
+  async getCustomField(id, defaultValue = null)
+  {
+    const { customFields } = this.props;
+
+    if (! customFields) {
+      return Promise.reject(new Error('this object does not support custom fields'))
+    }
+
+    return customFields.getField(id, defaultValue);
   }
 
   /**
@@ -25,6 +44,22 @@ class ContextObject
     }
 
     return this.props.customFields;
+  }
+
+  /**
+   * Retrieves a property from this object
+   *
+   * @param {String} [property]
+   * @return {Promise<*>}
+   */
+  async get(property)
+  {
+    if (property && typeof property !== 'string') {
+      return Promise.reject(new Error('property must be a string'));
+    }
+
+    const path = property ? property.split('.') : [];
+    return this.props.propertyProvider({ path, type: this.props.type });
   }
 
   /**